requirements.txt
prettytable==3.12.0
pytest==8.3.4
mysql-connector-python~=9.2.0